Fix overflows in intermediate calculations in sysctl_msec_to_ticks(). At hz values of 1000 and above the overflows caused net.inet.tcp.keepidle to be reported as negative.
MFC after:	3 days


git-svn-id: svn+ssh://svn.freebsd.org/base/head@100420 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/netinet/tcp_timer.c
@@ -74,13 +74,13 @@ sysctl_msec_to_ticks(SYSCTL_HANDLER_ARGS)
 	int error, s, tt;
 
 	tt = *(int *)oidp->oid_arg1;
-	s = tt * 1000 / hz;
+	s = (int)((int64_t)tt * 1000 / hz);
 
 	error = sysctl_handle_int(oidp, &s, 0, req);
 	if (error || !req->newptr)
 		return (error);
 
-	tt = s * hz / 1000;
+	tt = (int)((int64_t)s * hz / 1000);
 	if (tt < 1)
 		return (EINVAL);
